+++ pyobjc-framework-Cocoa/PyObjCTest/test_nshashtable.py
@@ -4,6 +4,48 @@ from Foundation import *
 
 
 class TestNSHashTable (TestCase):
+    def testConvenience(self):
+        v = NSHashTable.hashTableWithOptions_(NSPointerFunctionsObjectPersonality)
+        self.assertEqual(len(v), 0)
+
+        v.add(32)
+        v.add(2)
+
+        self.assertIn(32, v)
+        self.assertNotIn(42, v)
+
+        self.assertNotIn(1, v)
+        self.assertIn(2, v)
+        v.remove(2)
+        v.remove(1)
+        self.assertNotIn(1, v)
+        self.assertNotIn(2, v)
+
+        v.clear()
+        self.assertEqual(len(v), 0)
+
+        v.add(1)
+        v.add(2)
+        v.add(3)
+
+        self.assertEqual(len(v), 3)
+        self.assertIn(v.pop(), {1,2,3})
+        self.assertIn(v.pop(), {1,2,3})
+        self.assertIn(v.pop(), {1,2,3})
+        self.assertEqual(len(v), 0)
+
+        v.add(None)
+        self.assertEqual(len(v), 1)
+        self.assertIn(None, v)
+        v.remove(None)
+        self.assertEqual(len(v), 0)
+
+        v.add(None)
+        o = v.pop()
+        self.assertEqual(o, None)
+        self.assertEqual(len(v), 0)
+
+
     def testConstants(self):
         self.assertEqual(NSHashTableStrongMemory, 0)
         self.assertEqual(NSHashTableZeroingWeakMemory, NSPointerFunctionsZeroingWeakMemory)