+++ pyobjc-framework-Cocoa/PyObjCTest/test_nslocalizedstring.py
@@ -2,7 +2,7 @@
 from PyObjCTools.TestSupport import *
 import objc
 
-from Foundation import NSLocalizedString, NSAutoreleasePool
+from Foundation import NSLocalizedString, NSAutoreleasePool, NSLocalizedStringFromTable, NSLocalizedStringFromTableInBundle, NSLocalizedStringWithDefaultValue, NSBundle
 
 class TestNSLocalizedString(TestCase):
     def testBasic(self):
@@ -17,6 +17,22 @@ class TestNSLocalizedString(TestCase):
         self.assertEqual (s, b"hello world".decode('ascii'))
         # XXX : Since we get the same object back, it's still unicode
         #self.assertEqual (s.nsstring().description(), b"hello world".decode('ascii'))
+
+
+        pool = NSAutoreleasePool.alloc().init()
+        s = NSLocalizedStringFromTable(b"hello world".decode('ascii'), b"tab".decode('utf-8'), b"".decode('ascii'))
+        del pool
+        self.assertEqual (s, b"hello world".decode('ascii'))
+
+        pool = NSAutoreleasePool.alloc().init()
+        s = NSLocalizedStringFromTableInBundle(b"hello world".decode('ascii'), b"tab".decode('utf-8'), NSBundle.mainBundle(), b"".decode('ascii'))
+        del pool
+        self.assertEqual (s, b"hello world".decode('ascii'))
+
+        pool = NSAutoreleasePool.alloc().init()
+        s = NSLocalizedStringWithDefaultValue(b"hello world".decode('ascii'), b"tab".decode('utf-8'), NSBundle.mainBundle(), b"default".decode('utf-8'), b"".decode('ascii'))
+        del pool
+        self.assertEqual (s, b"default".decode('ascii'))
 
 if __name__ == '__main__':
     main( )